Refactor - 회원가입 컴포넌트 이름수정완료, 백그라운드 스타일 수정완료
