+++ src/pages/SignUp/SignUp.scss
@@ -1,21 +1,23 @@
-body {
-  background-color: #206648;
-}
-
 .signUp {
+  .backgroudBanner {
+    width: 100%;
+    background-color: #206648;
+    padding: 100px 0;
+  }
+
   .container {
-    margin: 200px;
+    margin: 0 170px;
     display: flex;
     flex-direction: column;
     background-color: white;
-    padding: 50px;
+    padding: 20px 50px;
   }
 
   .nav {
     font-size: 20px;
     border-bottom: #e9e9e9 2px solid;
-    margin-bottom: 50px;
-    padding-bottom: 35px;
+    margin-bottom: 40px;
+    padding-bottom: 20px;
     text-align: center;
   }
 
@@ -77,7 +79,7 @@ body {
     }
   }
   input {
-    margin: 10px;
+    margin: 8px;
   }
 
   .check {
